requirements.txt
# نظام إدارة ميكروتك - Khalid Soft
# MikroTik Management System Requirements

Flask==2.3.3
librouteros==3.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
requests==2.31.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.1
